chore: overall improvements on init/link commands
